058: update solution

+++ algorithm_0000-0999/058-Length of Last Word/main.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int lengthOfLastWord(string s) {
-        int rht = s.find_last_not_of(' ');
-        int lft = s.find_last_of(' ', rht);
-        lft = (lft == string::npos ? 0 : lft + 1);
+        size_t rht = s.find_last_not_of(' ');
+        size_t lft = s.find_last_of(' ', rht);
+        lft = lft == string::npos ? 0 : lft + 1;
         return rht - lft + 1;
     }
 };
